clients/python/bcdb/requirements.txt
grpcio==1.28.1
protobuf==3.10.0
PyNaCl==1.3.0
git+https://github.com/trezor/python-mnemonic
